+++ src/index.ts
@@ -2,6 +2,7 @@ export { doctorConfig, formatDoctor } from "./doctor";
 export { findConfigPath, loadConfig } from "./config";
 export { CONFIG_SCHEMA_URL } from "./constants";
 export { formatDecision } from "./format";
+export { collectGitHubUsage, formatGitHubUsageState, inferProvider, writeGitHubUsageState } from "./github-usage";
 export { starterConfig, writeStarterConfig } from "./init";
 export { formatPlan, planFreelane } from "./plan";
 export { formatProviderList, listProviders } from "./provider-list";